Voice Ai Agentic System/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic-settings==2.7.1

# Database
asyncpg==0.30.0
supabase==2.11.0

# AI/LLM
anthropic==0.42.0
groq==0.15.0

# Voice - STT
deepgram-sdk==3.8.0

# Voice - TTS
elevenlabs==1.13.5

# Voice - Telephony
twilio==9.4.0

# WhatsApp
httpx==0.28.1

# Web Search
tavily-python==0.5.0

# Hold Music
yt-dlp==2024.12.23

# Utilities
websockets>=12.0,<14.0
python-multipart==0.0.19
structlog==24.4.0
